requirements.txt
mediapipe>=0.10
opencv-python>=4.8
fastapi>=0.110
uvicorn>=0.29
httpx>=0.27
python-dotenv>=1.0
websockets>=12.0
